Increase max size for allocator
Increasing maximum allocatable size field from 16 bits to 31 bits.
Adding new test fixture for allocator using atomic mocks to cover hard to reach conditions.
Extend the documentation to account for field size.
Also includes a minor fix on the available size calculation.

GIT_ORIGIN_SPP_REV_ID: 0a2761db29f944d9c0fb92cf5b578d9f8733030e

+++ score/analysis/tracing/common/flexible_circular_allocator/lockless_flexible_circular_allocator.h
@@ -76,7 +76,7 @@ class LocklessFlexibleCircularAllocator : public IFlexibleCircularAllocator
                                              std::uint32_t aligned_size,
                                              std::uint32_t list_entry_element_index) noexcept;
     bool UpdateListEntryForAllocation(std::size_t list_entry_index,
-                                      std::uint16_t aligned_size,
+                                      std::uint32_t aligned_size,
                                       std::uint32_t offset) noexcept;
     score::Result<BufferBlock*> GetValidatedBlock(std::uint32_t offset) noexcept;
     bool ShouldResetBufferTail(std::uint32_t current_tail) const noexcept;
